+++ src/main/java/com/almostreliable/summoningrituals/SummoningRitualsClient.java
@@ -3,12 +3,14 @@ package com.almostreliable.summoningrituals;
 import com.almostreliable.summoningrituals.client.render.AltarRenderer;
 import com.almostreliable.summoningrituals.client.render.PatternPreviewRenderer;
 import com.almostreliable.summoningrituals.client.tooltip.PatternPreviewTooltipComponent;
+import com.almostreliable.summoningrituals.compat.kubejs.event.RitualRendererRegistryEvent;
 import com.almostreliable.summoningrituals.core.Registration;
 
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.fml.common.Mod;
 import net.neoforged.neoforge.client.event.EntityRenderersEvent;
+import net.neoforged.neoforge.client.event.RegisterClientReloadListenersEvent;
 import net.neoforged.neoforge.client.event.RegisterClientTooltipComponentFactoriesEvent;
 import net.neoforged.neoforge.client.event.RenderLevelStageEvent;
 import net.neoforged.neoforge.common.NeoForge;
@@ -17,17 +19,22 @@ import net.neoforged.neoforge.common.NeoForge;
 public final class SummoningRitualsClient {
 
     public SummoningRitualsClient(IEventBus eventBus) {
-        eventBus.addListener(SummoningRitualsClient::registerRenderers);
+        eventBus.addListener(SummoningRitualsClient::registerEntityRenderers);
         eventBus.addListener(SummoningRitualsClient::registerTooltipRenderers);
+        eventBus.addListener(SummoningRitualsClient::registerClientReloadListeners);
         NeoForge.EVENT_BUS.addListener(SummoningRitualsClient::onRenderLevelStage);
     }
 
-    private static void registerRenderers(EntityRenderersEvent.RegisterRenderers event) {
+    private static void registerEntityRenderers(EntityRenderersEvent.RegisterRenderers event) {
         event.registerBlockEntityRenderer(Registration.ALTAR_BLOCK_ENTITY.get(), AltarRenderer::new);
     }
 
     private static void registerTooltipRenderers(RegisterClientTooltipComponentFactoriesEvent event) {
         event.register(PatternPreviewTooltipComponent.Data.class, PatternPreviewTooltipComponent::new);
+    }
+
+    private static void registerClientReloadListeners(RegisterClientReloadListenersEvent event) {
+        event.registerReloadListener(new RitualRendererRegistryEvent.ReloadListener());
     }
 
     private static void onRenderLevelStage(RenderLevelStageEvent event) {