+++ src/main/java/com/almostreliable/summoningrituals/client/render/CustomRitualRenderer.java
@@ -1,0 +1,9 @@
+package com.almostreliable.summoningrituals.client.render;
+
+import com.almostreliable.summoningrituals.recipe.AltarRecipe;
+
+@FunctionalInterface
+public interface CustomRitualRenderer {
+
+    void render(AltarRenderer renderer, AltarRecipe recipe, AltarRenderContext context);
+}