+++ src/main/java/com/almostreliable/summoningrituals/compat/kubejs/event/package-info.java
@@ -1,0 +1,4 @@
+@NonnullDefault
+package com.almostreliable.summoningrituals.compat.kubejs.event;
+
+import org.lwjgl.system.NonnullDefault;