+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 
 - added client config option to disable candle particle rendering (useful for custom models)
+- added support for registering custom ritual renderers
 
 ## [3.6.0] - 2026-04-04